Fix rank party api to rank contorller
